refactor: centralize settings management with simpleStorage
- Create src/settings/state/appData.ts following the same pattern as tabs/state
- Migrate all storage operations to use simpleStorage utilities
- Maintain existing storage key structure (settings and version keys)
- Update Options page to use the new centralized settings management
- Remove old storage.ts file in favor of appData.ts
- Keep storage structure unchanged for backward compatibility
